*** empty log message ***
git-svn-id: https://svn.library.emory.edu/svn/slavevoyages/trunk@32 2440b244-9809-4ec7-bebd-234906d9abf5

+++ src/edu/emory/library/tas/dicts/Carib.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class Carib extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(3);
 	private static final String NAME = "Carib";
 	
 	public Carib() {
