+++ src/edu/emory/library/tas/dicts/Country.java
@@ -4,7 +4,7 @@ import edu.emory.library.tas.Dictionary;
 
 public class Country extends Dictionary {
 	
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(4);
 	private static final String NAME = "Country";
 	
 	public Country() {
