+++ src/edu/emory/library/tas/dicts/DeparturePort.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class DeparturePort extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(6);
 	private static final String NAME = "DeparturePort";
 	
 	public DeparturePort() {
