+++ src/edu/emory/library/tas/dicts/EmbRegion.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class EmbRegion extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(10);
 	private static final String NAME = "EmbRegion";
 	
 	public EmbRegion() {
