+++ src/edu/emory/library/tas/dicts/Fate2.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class Fate2 extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(11);
 	private static final String NAME = "Fate2";
 	
 	public Fate2() {
