+++ src/edu/emory/library/tas/dicts/Fate4.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class Fate4 extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(13);
 	private static final String NAME = "Fate4";
 	
 	public Fate4() {
