+++ src/edu/emory/library/tas/dicts/FirstDemPort.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class FirstDemPort extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(15);
 	private static final String NAME = "FirstDemPort";
 	
 	public FirstDemPort() {
