+++ src/edu/emory/library/tas/dicts/Flag2.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class Flag2 extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(16);
 	private static final String NAME = "Flag2";
 	
 	public Flag2() {
