+++ src/edu/emory/library/tas/dicts/ImpPrincipalRegion.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class ImpPrincipalRegion extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(18);
 	private static final String NAME = "ImpPrincipalRegion";
 	
 	public ImpPrincipalRegion() {
