+++ src/edu/emory/library/tas/dicts/MarkDegree.java
@@ -4,7 +4,7 @@ import edu.emory.library.tas.Dictionary;
 
 public class MarkDegree extends Dictionary {
 
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(22);
 	private static final String NAME = "MarkDegree";
 	
 	public MarkDegree() {
