+++ src/edu/emory/library/tas/dicts/MarkType.java
@@ -4,7 +4,7 @@ import edu.emory.library.tas.Dictionary;
 
 public class MarkType extends Dictionary {
 
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(25);
 	private static final String NAME = "MarkType";
 	
 	public MarkType() {
