+++ src/edu/emory/library/tas/dicts/Natinim3.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class Natinim3 extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(27);
 	private static final String NAME = "Natinim3";
 	
 	public Natinim3() {
