+++ src/edu/emory/library/tas/dicts/Nation.java
@@ -4,7 +4,7 @@ import edu.emory.library.tas.Dictionary;
 
 public class Nation extends Dictionary {
 
-	private static final Integer TYPE = new Integer(3);
+	private static final Integer TYPE = new Integer(29);
 	private static final String NAME = "Nation";
 	
 	public Nation() {
