+++ src/edu/emory/library/tas/dicts/OtherMark.java
@@ -4,7 +4,7 @@ import edu.emory.library.tas.Dictionary;
 
 public class OtherMark extends Dictionary {
 
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(33);
 	private static final String NAME = "OtherMark";
 	
 	public OtherMark() {
