+++ src/edu/emory/library/tas/dicts/PrimaryLast.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class PrimaryLast extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(37);
 	private static final String NAME = "PrimaryLast";
 	
 	public PrimaryLast() {
