+++ src/edu/emory/library/tas/dicts/SexAge.java
@@ -4,7 +4,7 @@ import edu.emory.library.tas.Dictionary;
 
 public class SexAge extends Dictionary {
 
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(42);
 	private static final String NAME = "SexAge";
 	
 	public SexAge() {
