+++ src/edu/emory/library/tas/dicts/Status.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class Status extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(45);
 	private static final String NAME = "Status";
 	
 	public Status() {
