+++ src/edu/emory/library/tas/dicts/Temp2.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class Temp2 extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(47);
 	private static final String NAME = "Temp2";
 	
 	public Temp2() {
