+++ src/edu/emory/library/tas/dicts/ThirdDemPort.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class ThirdDemPort extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(48);
 	private static final String NAME = "ThirdDemPort";
 	
 	public ThirdDemPort() {
