+++ src/edu/emory/library/tas/dicts/TonesType.java
@@ -3,7 +3,7 @@ package edu.emory.library.tas.dicts;
 import edu.emory.library.tas.Dictionary;
 
 public class TonesType extends Dictionary {
-	private static final Integer TYPE = new Integer(2);
+	private static final Integer TYPE = new Integer(49);
 	private static final String NAME = "TonesType";
 	
 	public TonesType() {
